backend/requirements.txt
flask
pymongo
flask-jwt-extended
flask-bcrypt
python-dotenv
fpdf
pillow
numpy